+++ app/formSchema/uiSchema/analyst/rfiUiSchema.ts
@@ -1,3 +1,5 @@
+import { EXCEL_FILE_EXTENSIONS } from '../constants';
+
 const rfiUiSchema = {
   'ui:order': [
     'rfiType',
@@ -423,6 +425,7 @@ export const rfiApplicantUiSchema = {
       'ui:widget': 'FileWidget',
       'ui:options': {
         hideOptional: true,
+        allowMultipleFiles: true,
       },
     },
     copiesOfRegistrationRfi: {
@@ -435,6 +438,7 @@ export const rfiApplicantUiSchema = {
       'ui:widget': 'FileWidget',
       'ui:options': {
         hideOptional: true,
+        allowMultipleFiles: true,
       },
     },
     preparedFinancialStatementsRfi: {
@@ -447,6 +451,9 @@ export const rfiApplicantUiSchema = {
       'ui:widget': 'FileWidget',
       'ui:options': {
         hideOptional: true,
+        allowMultipleFiles: true,
+        fileTypes:
+          '.pdf, .png, .jpg, .jpeg, .vsd, .vsdx, .doc, .docx, .ppt, .pptx',
       },
     },
     logicalNetworkDiagramRfi: {
@@ -459,6 +466,8 @@ export const rfiApplicantUiSchema = {
       'ui:widget': 'FileWidget',
       'ui:options': {
         hideOptional: true,
+        allowMultipleFiles: true,
+        fileTypes: `${EXCEL_FILE_EXTENSIONS}, .mpp`,
       },
     },
     projectScheduleRfi: {
@@ -471,6 +480,7 @@ export const rfiApplicantUiSchema = {
       'ui:widget': 'FileWidget',
       'ui:options': {
         hideOptional: true,
+        allowMultipleFiles: true,
       },
     },
     communityRuralDevelopmentBenefitsRfi: {
@@ -483,6 +493,7 @@ export const rfiApplicantUiSchema = {
       'ui:widget': 'FileWidget',
       'ui:options': {
         hideOptional: true,
+        allowMultipleFiles: true,
       },
     },
     otherSupportingMaterialsRfi: {
@@ -519,6 +530,8 @@ export const rfiApplicantUiSchema = {
       'ui:widget': 'FileWidget',
       'ui:options': {
         hideOptional: true,
+        allowMultipleFiles: true,
+        fileTypes: '.kml, .kmz',
       },
     },
     currentNetworkInfastructureRfi: {
@@ -531,6 +544,8 @@ export const rfiApplicantUiSchema = {
       'ui:widget': 'FileWidget',
       'ui:options': {
         hideOptional: true,
+        allowMultipleFiles: true,
+        fileTypes: '.kml, .kmz',
       },
     },
     upgradedNetworkInfrastructureRfi: {