+++ app/lib/relay/index.ts
@@ -1,1 +1,0 @@
-export { default as relayEnvironment } from './relayEnvironment';