+++ db/test/unit/mutations/create_assessment_form_test.sql
@@ -19,6 +19,9 @@ values('2022-03-01 09:00:00-07', '2022-05-01 09:00:00-07', 1);
 
 select mocks.set_mocked_time_in_transaction('2022-04-01 09:00:00-07'::timestamptz);
 set jwt.claims.sub to 'testCcbcAuthUser';
+insert into ccbc_public.ccbc_user
+  (given_name, family_name, email_address, session_sub) values
+  ('foo1', 'bar', 'foo1@bar.com', 'testCcbcAuthUser');
 set role ccbc_auth_user;
 
 -- Set role to job_executor and put in test slug for assessment forms