+++ db/test/unit/tables/application_analyst_lead_test.sql
@@ -8,8 +8,13 @@ select has_column('ccbc_public', 'application_analyst_lead', 'analyst_id', 'tabl
 
 select mocks.set_mocked_time_in_transaction('2022-04-01 09:00:00-07'::timestamptz);
 
-insert into ccbc_public.intake(open_timestamp, close_timestamp, ccbc_intake_number)
-values('2022-03-01 09:00:00-07', '2022-05-01 09:00:00-07', 1);
+insert into
+  ccbc_public.intake(id, open_timestamp, close_timestamp, ccbc_intake_number)
+overriding system value
+values
+  (1, '2022-03-01 09:00:00-07', '2022-05-01 09:00:00-07', 1);
+
+delete from ccbc_public.ccbc_user;
 
 -- ccbc_guest
 set role ccbc_guest;
@@ -40,9 +45,18 @@ select throws_like(
 
 -- ccbc_auth_user
 set role ccbc_auth_user;
-set jwt.claims.sub to '11111111-1111-1111-1111-111111111111';
+set jwt.claims.sub to 'testCcbcAuthUser';
 
-select ccbc_public.create_application();
+insert into ccbc_public.ccbc_user
+  (id, given_name, family_name, email_address, session_sub) overriding system value values
+  (1, 'foo1', 'bar', 'foo1@bar.com', 'testCcbcAuthUser');
+
+-- select ccbc_public.create_application();
+
+insert into ccbc_public.application
+  (id, ccbc_number, owner, intake_id) overriding system value
+   values
+  (1,'CCBC-010001', 'testCcbcAuthUser', 1);
 
 select throws_like(
   $$
@@ -72,8 +86,12 @@ reset role;
 
 -- ccbc_admin
 set role ccbc_admin;
-set jwt.claims.sub to '11111111-1111-1111-1111-111111111111';
 
+set jwt.claims.sub to '11111111-1111-1111-1111-111111111111';
+insert into ccbc_public.ccbc_user
+  (given_name, family_name, email_address, session_sub) values
+  ('foo2', 'bar', 'foo2@bar.com', '11111111-1111-1111-1111-111111111111');
+ 
 insert into ccbc_public.application_analyst_lead
   (application_id, analyst_id) overriding system value
   values