+++ db/test/unit/tables/assessment_data_test.sql
@@ -24,7 +24,13 @@ values('2022-03-01 09:00:00-07', '2022-05-01 09:00:00-07', 1);
 
 select mocks.set_mocked_time_in_transaction('2022-04-01 09:00:00-07'::timestamptz);
 set jwt.claims.sub to 'testCcbcAuthUser';
-
+insert into ccbc_public.ccbc_user
+  (given_name, family_name, email_address, session_sub) values
+  ('foo1', 'bar', 'foo1@bar.com', 'testCcbcAuthUser');
+insert into ccbc_public.ccbc_user
+  (given_name, family_name, email_address, session_sub) values
+  ('foo2', 'bar', 'foo1@bar.com', '11111111-1111-1111-1111-111111111111');
+  
 set role ccbc_auth_user;
 
 select ccbc_public.create_application();